verifyFeur support multi-word (verifyWord now) add "c ki" event
